Merge in PNM/ws23---pa---nm---gps-basierte-stolperstein-ar-informationsanzeige from feature/WS23PASTST-28/stumbling-block-view-in-fe to master

* commit '4ea9153b7527ea4d29f997e7676d9813f55b0ae0':
  (WS23PASTST-28): Extend FE Functions
  (WS23PASTST-28): Extend Backend
  (WS23PASTST-28): Update Backend-Server and base implementation logic
  (WS23PASTST-28): Add Stumbling Block View to FE with initial load of all Stumbling Blocks
  install google maps support
  Update Angular to 17
  Update Angular to 17
  (FE): Add new Component Stolpersteine
  (Backend): Add Implementation for getting all Stumbling Blocks

+++ backend/src/main/java/com/thws/ar/stolpersteine/backend/db/entity/Address.java
@@ -12,7 +12,7 @@ public class Address {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer address_id;
+    private Long address_id;
 
     @Column(name = "street_name")
     private String streetName;
